+++ src/main/java/net/bios/frenchvanilla/Perms.java
@@ -6,4 +6,5 @@ public class Perms {
     }
 
     public static String CONFIG = permission("config");
+    public static String TELEPORT = permission("teleport");
 }