+++ src/main/java/net/bios/frenchvanilla/util/Pair.java
@@ -1,0 +1,11 @@
+package net.bios.frenchvanilla.util;
+
+public final class Pair<T, U> {
+    public final T left;
+    public final U right;
+
+    public Pair(T left, U right) {
+        this.left = left;
+        this.right = right;
+    }
+}